provider/maas: conditionally run ifdown/up
If the nett result of bridging interfaces results in nothing-to-do,
because they are already bridged, then don't run ifdown/up
unnecessarily.

This state can exist in MAAS 2.1 because you now have the option of
bridging interfaces ahead of deployment or bootstrap. And as we move
to dynamic bridging in Juju we want to invoke this script on an
interface on an ad-hoc basis but don't want it to do anything if that
interface is already bridged.

QA steps:

On a machine run:

 $ sudo ./add-juju-bridge.py /etc/network/interfaces \
     --activate \
     --interfaces-to-bridge=enp1s0f2

And note that enp1s0f2 gets bridged.

Verify with:

 $ brctl show

Run the add-juju-bridge.py script again with the same arguments and
now you'll see no action taken as it is already bridged as
br-enp1s0f2.

I also configured a node in MAAS 2.1 that had two bridged NICs;
bootstrapping a node doesn't additionally run ifdown/up because
there's nothing to do.

+++ provider/maas/add-juju-bridge.py
@@ -8,12 +8,20 @@
 #
 
 from __future__ import print_function
+
 import argparse
 import os
 import re
 import shutil
 import subprocess
 import sys
+
+# StringIO: accommodate Python2 & Python3
+
+try:
+    from StringIO import StringIO
+except ImportError:
+    from io import StringIO
 
 # These options are to be removed from a sub-interface and applied to
 # the new bridged interface.
@@ -368,7 +376,10 @@ def shell_cmd(s):
     return [out, err, p.returncode]
 
 
-def print_shell_cmd(s, verbose=True, exit_on_error=False):
+def print_shell_cmd(s, verbose=True, exit_on_error=False, dryrun=False):
+    if dryrun:
+        print(s)
+        return
     if verbose:
         print(s)
     out, err, retcode = shell_cmd(s)
@@ -380,22 +391,15 @@ def print_shell_cmd(s, verbose=True, exit_on_error=False):
         exit(1)
 
 
-def check_shell_cmd(s, verbose=False):
-    if verbose:
-        print(s)
-    output = subprocess.check_output(s, shell=True, stderr=subprocess.STDOUT).strip().decode("utf-8")
-    if verbose:
-        print(output.rstrip('\n'))
-    return output
-
-
 def arg_parser():
     parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter)
     parser.add_argument('--bridge-prefix', help="bridge prefix", type=str, required=False, default='br-')
     parser.add_argument('--one-time-backup', help='A one time backup of filename', action='store_true', default=True, required=False)
     parser.add_argument('--activate', help='activate new configuration', action='store_true', default=False, required=False)
     parser.add_argument('--interfaces-to-bridge', help="interfaces to bridge; space delimited", type=str, required=True)
+    parser.add_argument('--dryrun', help="dry run, no activation", action='store_true', default=False, required=False)
     parser.add_argument('--bridge-name', help="bridge name", type=str, required=False)
+    parser.add_argument('--bond-sleep-duration', help="duration to sleep between ifdown/up", type=int, required=False)
     parser.add_argument('filename', help="interfaces(5) based filename")
     return parser
 
@@ -418,7 +422,17 @@ def main(args):
         print_stanzas(stanzas)
         exit(0)
 
-    if args.one_time_backup:
+    # Dump stanzas to cur/new in-memory strings
+    cur = StringIO()
+    new = StringIO()
+    print_stanzas(stanzas, new)
+    print_stanzas(parser.stanzas(), cur)
+
+    if cur.getvalue() == new.getvalue():
+        print("already bridged, or nothing to do.")
+        exit(0)
+
+    if not args.dryrun and args.one_time_backup:
         backup_file = "{}-before-add-juju-bridge".format(args.filename)
         if not os.path.isfile(backup_file):
             shutil.copy2(args.filename, backup_file)
@@ -426,15 +440,16 @@ def main(args):
     ifquery = "$(ifquery --interfaces={} --exclude=lo --list)".format(args.filename)
 
     print("**** Original configuration")
-    print_shell_cmd("cat {}".format(args.filename))
-    print_shell_cmd("ifconfig -a")
-    print_shell_cmd("ifdown --exclude=lo --interfaces={} {}".format(args.filename, ifquery))
+    print_shell_cmd("cat {}".format(args.filename), dryrun=args.dryrun)
+    print_shell_cmd("ifconfig -a", dryrun=args.dryrun)
+    print_shell_cmd("ifdown --exclude=lo --interfaces={} {}".format(args.filename, ifquery), dryrun=args.dryrun)
 
     print("**** Activating new configuration")
 
-    with open(args.filename, 'w') as f:
-        print_stanzas(stanzas, f)
-        f.close()
+    if not args.dryrun:
+        with open(args.filename, 'w') as f:
+            print_stanzas(stanzas, f)
+            f.close()
 
     # On configurations that have bonds in 802.3ad mode there is a
     # race condition betweeen an immediate ifdown then ifup.
@@ -448,15 +463,15 @@ def main(args):
         if s.is_logical_interface and s.iface.is_bonded:
             print("working around https://bugs.launchpad.net/ubuntu/+source/ifenslave/+bug/1269921")
             print("working around https://bugs.launchpad.net/juju-core/+bug/1594855")
-            print_shell_cmd("sleep 3")
+            print_shell_cmd("sleep 3", dryrun=args.dryrun)
             break
 
-    print_shell_cmd("cat {}".format(args.filename))
-    print_shell_cmd("ifup --exclude=lo --interfaces={} {}".format(args.filename, ifquery))
-    print_shell_cmd("ip link show up")
-    print_shell_cmd("ifconfig -a")
-    print_shell_cmd("ip route show")
-    print_shell_cmd("brctl show")
+    print_shell_cmd("cat {}".format(args.filename), dryrun=args.dryrun)
+    print_shell_cmd("ifup --exclude=lo --interfaces={} {}".format(args.filename, ifquery), dryrun=args.dryrun)
+    print_shell_cmd("ip link show up", dryrun=args.dryrun)
+    print_shell_cmd("ifconfig -a", dryrun=args.dryrun)
+    print_shell_cmd("ip route show", dryrun=args.dryrun)
+    print_shell_cmd("brctl show", dryrun=args.dryrun)
 
 # This script re-renders an interfaces(5) file to add a bridge to
 # either all active interfaces, or a specific interface.
